add symlink
I'm adding this symlink to make github pages serve the right file, quiplibre.html, from this directory without renaming the file to the unhelpful "index.html".

This worked, so I deleted the gh-pages branch I had on github which had quiplibre.html renamed to index.html. I had to keep manually updating that to rename the file, so it was not worth the trouble it took to maintain!